Moves DelayedQuery{T} form Internals
DelayedScalarQuery{T} is in Xtensive.Orm but
DelayedQuery{T} is in Xtensive.Orm.Internals.
Both types are directly accesible.
